Created gamify.py, added main tables.
Added 'user_points', 'achievements', 'images' and 'unlockables' tables.

'user_points': Keeps track of the points earned by the users. 
'achievements': Keeps a list of achievements and their thresholds.
'images': Keeps the path of the images of users, if the profile avatar feature is unlocked.
'unlockables': Keeps a list of features and whether or not they are unlocked for each user.