Added functions fill the tables automatically
Added make_gamify_entry and fill_achievements_table functions.

make_gamify_entry: This function is called whenever a new entry is made in the auth_user table. All tables are filled with their default information for the new user.

fill_achievements_table : This function fills the achievements table with a list of achievements and their thresholds. It's run once when the tables are being built.