Added modifyUserPoints function
Added the new function modifyUserPoints. This function is called via jQuery and takes an integer variable named pointDelta as the parameter data. This value is then added to the points field in the 'user_points' table. If the value is negative 'all_time_points' field is not modified but if the value is positive 'all_time_points' field is modified together with the 'points' field.