Added unlockFeature function
Added the new function unlockFeature. This function is called via jQuery and takes a string variable named feature as the parameter data. This string is then matched with one of the existing feature names and the matching feature is unlocked in the 'unlockables' table. Unlocking is done by simply changing the integer value 0 to 1.