Added getUnlockedFeatures function
Added the new function getUnlockedFeatures. This function is called via jQuery and it does not take any parameters. It returns the record of the user in the 'unlockables' table in an list with 1 element and it's returned as json. On the other end user should parse this list and access the only element in it using the first() function. If a feature has the value 1 it is unlocked, if it has the value 0 then it is still locked.