Modified index function.
Added two new tables to the returning dict in the index function. Points table holds the record that corresponds to the current user id. Achievements table holds the entire list of achievements, one record for each achievement. These tables are used to display the leaderboard and the earned achievements in the "Progress Page".