Added shop function
Added the new function shop. This function is called when the newly added "Shop Page" is being loaded. Function handles authentication. Then it gets the information for the current student and returns it in the returning dict. It also returns students current points in the variable points. It later returns the unlockable feature list for the student in the unlockables variable. In the "Shop Page" points value is displayed and a list of features are shown. Features are grayed out if they are not affordable or if they are already unlocked.