Created shop.html, used when "Shop Page" is loaded
Created shop.html, used when "Shop Page" is loaded. The page shows the student name and points. Then shows a list of features and buttons next to them to buy them. Buttons have the costs written on them. If the user can't afford a feature or the feature is already bought then the button is grayed out. This page requires the newly added shop function in assignments.py to work at all.