Updated user.html to handle profile avatar feature
If the user hasn't bought the profile avatar feature then the page is displayed normally, allowImage = 0. 

If the user has bought the profile avatar feature but hasn't uploaded an image yet a form is displayed to upload images, allowImage = 1. This removes the information on the right side of the page to make room for the form.

If the user has bought the profile avatar feature and uploaded an image then that image will be displayed along with a button to remove the image, allowImage = 2. This removes the information on the right side of the page to make room for the image.